Huge messy refactor
This is refactoring without backward compat. Won't be able to migrate
existing dbs (for now). (extra effort)

We can now delete an item at the correct position (and we do not just filter
according to its name). This required us to have an extra intermediary table
in an django M2M relation that saves each product-order relation.

Also, we do not pass around weird json when adding and removing things. Adding
happens by passing a product.id and removing by passing a "ProductInOrder" id.

The new removal API has the advantage that we can remove a product for
multiple open orders as one of the goals was to have multiple orders open
simultanously.

Everything should still work