Add basic front-end only change calculator
See issue #75
This commit adds a modal on the front-end that displays when paying by
cash, it grabs the total amount via ajax and calculates required change
based on input.

The ajax request is stupid. Utterly awful. The front-end already knows
about the total amount, and I needed to add an API point to grab it,
just for it to be usable in this dialog.